% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern_palettes.R
\name{modern_palette}
\alias{modern_palette}
\title{modern palette generator}
\usage{
modern_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{pgl}}

\item{n}{Number of colors desired. If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colors.}
}
\value{
A vector of colors.
}
\description{
These are a handful of color palettes from the PGL2024
}
\examples{
modern_palette("pgl")

# If you need more colors than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colors
pal <- modern_palette(name = "pgl", n = 2, type = "continuous")
}
\keyword{colors}
