% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modern_palettes.R
\docType{data}
\name{modern_palettes}
\alias{modern_palettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 1.
}
\usage{
modern_palettes
}
\description{
Use names(modern) to return all possible palette names. Current choices are:
\code{pgl}
}
\keyword{datasets}
